+++ src/appl/strsignif.c
@@ -113,8 +113,9 @@ void str_signif(char *x, int *n, const char **type, int *width, int *digits,
     double xx;
     int iex, j, len_flag = strlen(*flag);
 
-    char *f0  =	 R_alloc(do_fg ? 1+1+len_flag+3 : 1, sizeof(char));
-    char *form = R_alloc(1+1+len_flag+3 + strlen(*format), sizeof(char));
+    char *f0  =	 R_alloc((size_t) do_fg ? 1+1+len_flag+3 : 1, sizeof(char));
+    char *form = R_alloc((size_t) 1+1+len_flag+3 + strlen(*format),
+			 sizeof(char));
 
     if (wid == 0)
 	error(_(".C(..): Width cannot be zero"));
