+++ src/nmath/signrank.c
@@ -72,7 +72,7 @@ w_init_maybe(int n)
     }
 
     if(!w) {
-	w = (double *) calloc(c + 1, sizeof(double));
+	w = (double *) calloc((size_t) c + 1, sizeof(double));
 #ifdef MATHLIB_STANDALONE
 	if (!w) MATHLIB_ERROR("%s", _("signrank allocation error"));
 #endif
